Fix firmware cache key to account for erisc_iram toggling (#27810) (#29986)
### Ticket
Fixes #27810
Link to Github Issue:
https://github.com/tenstorrent/tt-metal/issues/27810

### Problem description
Resolves the hang in `TestUnicastConnAPI2DSmoke` that occurred when
Fabric2DFixture tests ran after MeshDeviceFixture or UnitMeshCQFixture
tests.

**Root Cause:**
The firmware cache tracking key (`MetalContext::firmware_built_keys_`)
only used `build_key`, which didn't capture firmware compilation setting
changes. When `MetalContext::set_fabric_config()` toggled
`rtoptions.erisc_iram_enabled` mid-process (between test fixtures), the
firmware cache incorrectly considered the firmware already built and
reused the cached firmware compiled with the wrong linker script. This
caused:

1. First few tests: ERISC firmware compiled without `ENABLE_IRAM` macro
(erisc_iram disabled)
2. TestUnicastConnAPI2DSmoke: runtime enabled erisc_iram, but
`build_key` unchanged -> already cached firmware with wrong linker
script was used (runtime/hw/toolchain/wormhole/erisc-b0-app.ld instead
of runtime/hw/toolchain/wormhole/erisc-b0-app_iram.ld)
3. Result: Wrong firmware causes ethernet cores to become inaccessible
via NoC-reads -> fabric hangs.

The firmware path remained the same for both tests because only
`build_key` was used.


### What's changed

#### 1. Fix Firmware Cache Tracking (Core Fix)
Changed firmware cache key to use full 64-bit `fw_compile_hash` combined
with `build_key` using XOR.
This ensures the firmware cache correctly detects when firmware
compilation settings change and rebuilds ERISC firmware with the
appropriate linker script and `ENABLE_IRAM` macro.

- `metal_context.cpp/hpp` : Changed `fw_build_key` to use XOR combining
: `(build_key ^ fw_compile_hash)`, updated `firmware_built_keys_` from
`uint32_t` to `uint64_t`
- `jit_build/build.cpp/hpp` : Added `fw_compile_hash` parameter to
`JitBuildEnv::init()`, firmware path now includes both:
`.../build_key/fw_compile_hash/firmware/`
- `build_env_manager.cpp` : Passes `fw_compile_hash` (computed from
`rtoptions().get_compile_hash_string()`) to `JitBuildEnv::init`


- Use `std::filesystem::path` to compose `out_firmware_root_` and
`out_kernel_root_`



#### 2. Add Inspector RPC for Device-Specific Firmware Path Resolution
(Required Support)

Since the firmware paths now include `fw_compile_hash` in the directory
structure, triage tools can no longer use hardcoded relative paths.

To avoid hardcoded relative paths and support multiple device/build
variants, added an Inspector RPC interface to dynamically query
device-specific firmware paths.

Previously, `dispatcher_data.py` constructed firmware paths using
relative paths (e.g., `../../../firmware/brisc/brisc.elf`), which breaks
with the new path structure and doesn't work when build configuration
changes result in different firmware locations.

- Centralized RPC callback registration in
`tt::tt_metal::inspector::Data`; call
`Inspector::set_build_env_fw_compile_hash(...)` from
`MetalContext::initialize()`
- Added `BuildEnvManager::get_all_build_envs_info()` returning
`std::vector<BuildEnvInfo>` copies with proper mutex locking to avoid
race conditions with concurrent `add_build_env()` calls during
initialization
- Inspector RPC now exposes only `getAllBuildEnvs()` returning all
device build environments upfront (removed per-device `getBuildEnv()` to
prevent lazy build env creation)
- Updated Inspector RPC interface (`rpc.capnp`): removed `getBuildEnv()`
method, kept only `getAllBuildEnvs()` returning
`List(BuildEnvPerDevice)` with `BuildEnvData` structs
- RPC thread-safety: `fw_compile_hash` stored as `std::atomic<uint64_t>`
in tt::tt_metal::inspector::Data
- Updated `dispatcher_data.py` to query firmware paths via RPC instead
of constructing relative paths; prefill cache via `getAllBuildEnvs()`\
- Python side: prefills `_build_env_cache` at initialization, requires
`RunChecks.devices` with no fallback logic, fails fast with clear errors
if device not in cache; removed unused `get_firmware_elf_path()` and
`_a_kernel_path`

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable)
- [ ] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml)
CI passes (if applicable) See [recommended dev
flow](https://github.com/tenstorrent/tt-metal/blob/main/models/docs/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [ ] [Galaxy
quick](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-quick.yaml)
CI passes (if applicable)
- [ ] [Galaxy demo tests, for
Llama](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-demo-tests.yaml)
CI passes, if applicable, because of current Llama work
- [ ] (For runtime and ops writers) [T3000 unit
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml)
CI passes (if applicable, since this is run on push to main)
- [ ] (For models and ops writers) [T3000 demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml)
CI passes (if applicable, since this is required for release)
- [ ] New/Existing tests provide coverage for changes

+++ tt_metal/impl/context/metal_context.hpp
@@ -150,7 +150,7 @@ private:
     size_t fw_compile_hash_ = 0;  // To check if FW recompilation is needed
 
     // Used to track which FW has been built already
-    std::unordered_set<uint32_t> firmware_built_keys_;
+    std::unordered_set<uint64_t> firmware_built_keys_;
 
     // Written to device as part of FW init, device-specific
     std::unordered_map<chip_id_t, std::vector<int32_t>> dram_bank_offset_map_;
